requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.10.0
mysqlclient==2.1.1
platformdirs==3.1.1
pytz==2022.7.1
sqlparse==0.4.3
virtualenv==20.21.0
